working, not working yet

+++ pg.plugin.zsh
@@ -1,17 +1,19 @@
 #!/bin/zsh
 
 __pg-check() {
-  if [ -z "$2" ]; then
+  if [ "$#" = 2 ]; then
     echo "Usage: pg $1 <db_name>"
-    return 0
+    return 1
   fi
+  return 0
 }
 
 __pg-check2() {
-  if [ -z "$2" ] || [ -z "$3" ]; then
+  if [ "$#" = 3 ]; then
     echo "Usage: pg $1 <origin> <target>"
-    return 0
+    return 1
   fi
+  return 0
 }
 
 _pg-ls() {
@@ -20,46 +22,59 @@ _pg-ls() {
 
 _pg-kill-connections() {
   local db_name="$*"
-  __pg-check "kill-connections" "$db_name"
-  psql postgres > /dev/null <<EOF
+  if __pg-check "kill-connections" "$db_name"; then
+    psql postgres > /dev/null <<EOF
     SELECT pg_terminate_backend(pid)
     FROM pg_stat_activity
     WHERE pid <> pg_backend_pid()
     AND datname='$db_name';
 EOF
+  fi
 }
 
 _pg-create() {
   local db_name="$*"
-  __pg-check "create" "$db_name"
-  createdb "$db_name"
+  __pg-check "create" "$db_name" && \
+    createdb "$db_name"
 }
 
 _pg-drop() {
   local db_name="$*"
-  __pg-check "drop" "$db_name"
-  _pg-kill-connections "$db_name"
-  dropdb --if-exists "$db_name"
+  if __pg-check "drop" "$db_name"; then
+    _pg-kill-connections "$db_name"
+    dropdb --if-exists "$db_name"
+  fi
 }
 
 _pg-cp() {
   local origin="$1"
   local target="$2"
-  __pg-check2 "cp" "$origin" "$target"
-  _pg-kill-connections "$origin"
-  psql postgres &> /dev/null <<EOF
-    CREATE DATABASE "$target" WITH TEMPLATE "$origin";
+  if __pg-check2 "cp" "$origin" "$target"; then
+    _pg-kill-connections "$origin"
+    psql postgres &> /dev/null <<EOF
+      CREATE DATABASE "$target" WITH TEMPLATE "$origin";
 EOF
+  fi
 }
 
 _pg-mv() {
   local origin="$1"
   local target="$2"
-  __pg-check2 "mv" "$origin" "$target"
-  _pg-kill-connections "$origin"
-  psql postgres &> /dev/null <<EOF
-    ALTER DATABASE "$origin" RENAME TO "$target";
+  if __pg-check2 "mv" "$origin" "$target"; then
+    _pg-kill-connections "$origin"
+    psql postgres &> /dev/null <<EOF
+      ALTER DATABASE "$origin" RENAME TO "$target";
 EOF
+  fi
+}
+
+_pg-dump-table() {
+  local db_name="$1"
+  local table_name="$2"
+  if [ "$#" != 2 ]; then
+    echo "Usage: pg dump-table <db_name> <table_name>"
+    pg_dump --table="$table_name" --data-only --column-inserts "$db_name"
+  fi
 }
 
 pg() {
@@ -83,8 +98,11 @@ pg() {
     mv)
       _pg-mv "$@"
       ;;
+    dump-table)
+      _pg-dump-table "$@"
+      ;;
     *)
-      echo "Usage: pg (ls|kill-connections|create|drop|cp|mv) <args>"
+      echo "Usage: pg (ls|kill-connections|create|drop|cp|mv|dump-table) <args>"
       return 0
       ;;
   esac
